Refactor: event -> kafka 전환, common 모듈 정리

+++ product/src/main/java/dukku/product/boundedContext/product/app/cqrs/ProductSyncFacade.java
@@ -15,7 +15,7 @@ public class ProductSyncFacade {
         syncProductStatsUseCase.execute();
     }
 
-    public void syncProductToElasticsearch(int productId) {
+    public void syncProductToElasticsearch(Long productId) {
         deleteProductSyncUseCase.deletedProductToElasticsearch(productId);
     }
 }
